requirements.txt
django==1.9.*
djangorestframework==3.3.*
